refactor: move mp3 frame parsing into class
